fix: rename common components
